spark/requirements_test.txt
# Requirements for testing with GitHub workflows.

--extra-index-url https://download.pytorch.org/whl/cpu
filelock==3.12.3
Jinja2==3.1.2
kornia==0.7.0
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
Pillow==9.0.1
pytest==7.4.0
scipy==1.11.1
sympy==1.12
torch==2.0.1+cpu
typing_extensions==4.7.1
